gl: Refactor the rest of GLHelpers
