+++ rpcs3/Emu/RSX/GL/GLFragmentProgram.h
@@ -2,6 +2,7 @@
 #include "../Program/FragmentProgramDecompiler.h"
 #include "../Program/GLSLTypes.h"
 #include "GLHelpers.h"
+#include "glutils/program.h"
 
 namespace glsl
 {
