+++ rpcs3/Emu/RSX/GL/GLShaderInterpreter.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "GLHelpers.h"
+#include "glutils/program.h"
 #include "../Program/ProgramStateCache.h"
 #include "../Common/TextureUtils.h"
