+++ rpcs3/GLGSRender.vcxproj
@@ -60,13 +60,18 @@
     <ClInclude Include="Emu\RSX\GL\GLGSRender.h" />
     <ClInclude Include="Emu\RSX\GL\GLProcTable.h" />
     <ClInclude Include="Emu\RSX\GL\GLProgramBuffer.h" />
+    <ClInclude Include="Emu\RSX\GL\glutils\blitter.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\buffer_object.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\capabilities.hpp" />
     <ClInclude Include="Emu\RSX\GL\glutils\common.h" />
+    <ClInclude Include="Emu\RSX\GL\glutils\fbo.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\pixel_settings.hpp" />
+    <ClInclude Include="Emu\RSX\GL\glutils\program.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\ring_buffer.h" />
     <ClInclude Include="Emu\RSX\GL\glutils\state_tracker.hpp" />
     <ClInclude Include="Emu\RSX\GL\glutils\image.h" />
+    <ClInclude Include="Emu\RSX\GL\glutils\sync.hpp" />
+    <ClInclude Include="Emu\RSX\GL\glutils\vao.hpp" />
     <ClInclude Include="Emu\RSX\GL\GLVertexProgram.h" />
     <ClInclude Include="Emu\RSX\GL\GLHelpers.h" />
     <ClInclude Include="Emu\RSX\GL\GLRenderTargets.h" />
@@ -83,8 +88,11 @@
     <ClCompile Include="Emu\RSX\GL\GLGSRender.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLOverlays.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLPipelineCompiler.cpp" />
+    <ClCompile Include="Emu\RSX\GL\glutils\blitter.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\buffer_object.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\common.cpp" />
+    <ClCompile Include="Emu\RSX\GL\glutils\fbo.cpp" />
+    <ClCompile Include="Emu\RSX\GL\glutils\program.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\ring_buffer.cpp" />
     <ClCompile Include="Emu\RSX\GL\glutils\image.cpp" />
     <ClCompile Include="Emu\RSX\GL\GLVertexProgram.cpp" />
